+++ lib/src/containers/list.dart
@@ -11,50 +11,48 @@ class ArnaList extends StatelessWidget {
   /// The items of list view.
   final List<Widget> items;
 
-  Widget _buildChild(BuildContext context) {
-    return ConstrainedBox(
-      constraints: BoxConstraints(
-        maxWidth:
-            expanded(context) ? deviceWidth(context) * 0.56 : double.infinity,
-      ),
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: items,
-      ),
-    );
-  }
-
   @override
   Widget build(BuildContext context) {
     return Padding(
       padding: Styles.listPadding,
       child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
         children: <Widget>[
-          if (title != null)
-            Padding(
-              padding: EdgeInsets.fromLTRB(
-                expanded(context)
-                    ? deviceWidth(context) * 0.12 + Styles.padding
-                    : Styles.padding,
-                Styles.padding,
-                expanded(context)
-                    ? deviceWidth(context) * 0.12 + Styles.padding
-                    : Styles.padding,
-                Styles.padding,
-              ),
-              child: Row(
-                children: <Widget>[
-                  Flexible(
-                    child: Text(
-                      title!,
-                      style: ArnaTheme.of(context).textTheme.textStyle,
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          _buildChild(context),
+          const SizedBox(height: 1, width: double.infinity),
+          LayoutBuilder(
+            builder: (BuildContext context, BoxConstraints constraints) {
+              return ConstrainedBox(
+                constraints: BoxConstraints(
+                  maxWidth: expanded(context)
+                      ? constraints.maxWidth > Styles.expanded
+                          ? constraints.maxWidth * 0.7
+                          : double.infinity
+                      : double.infinity,
+                ),
+                child: Column(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  mainAxisSize: MainAxisSize.min,
+                  children: <Widget>[
+                    if (title != null)
+                      Padding(
+                        padding: Styles.normal,
+                        child: Row(
+                          children: <Widget>[
+                            Flexible(
+                              child: Text(
+                                title!,
+                                style:
+                                    ArnaTheme.of(context).textTheme.textStyle,
+                              ),
+                            ),
+                          ],
+                        ),
+                      ),
+                    ...items,
+                  ],
+                ),
+              );
+            },
+          ),
         ],
       ),
     );