+++ lib/src/inputs/segmented_control.dart
@@ -268,7 +268,6 @@ class _ArnaSegmentedControlItem extends StatelessWidget {
         );
       },
       onPressed: onPressed,
-      tooltipMessage: label,
       cursor: cursor,
     );
   }